修复menu 图片文字错误 添加title(看需不需要修改成toolbar形式) 日志查询 时间倒序 添加输入Log文本界面 未完成

+++ app/src/main/java/com/aaa/cd/util/Utils.java
@@ -7,9 +7,9 @@ import java.text.DecimalFormat;
  * Created by aaa on 2016/9/4.
  */
 public class Utils {
-    public static String getDuration(int timeMillis){
+    public static String getDuration(long timeMillis){
 
-        int seconds=timeMillis/1000;
+        long seconds=timeMillis/1000;
         if(seconds>=86400)
         {
             return seconds/86400+"天"+(seconds%86400)/3600+"小时";
